ui: refactor selectors usage for IndexDetails page
Before, we had single selector that included everything needed for
IndexDetails page that is not effective as far as it'll be recomputed all
props with any small change.
Current change reuses almost the same logic as before but breaks it down to
be a part of `mapStateToProps` props that are more independent from each other.

Release note: None
